tests/wpt/tests/css/css-overscroll-behavior/inheritance.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>Inheritance of CSS Overscroll Behavior properties</title>
<link rel="help" href="https://drafts.csswg.org/css-overscroll-behavior/#property-index">
<meta name="assert" content="Properties inherit or not according to the spec.">
<meta name="assert" content="Properties have initial values according to the spec.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/inheritance-testcommon.js"></script>
</head>
<body>
<div id="container">
  <div id="target"></div>
</div>
<script>
assert_not_inherited('overscroll-behavior-block', 'auto', 'contain');
assert_not_inherited('overscroll-behavior-inline', 'auto', 'contain');
assert_not_inherited('overscroll-behavior-x', 'auto', 'contain');
assert_not_inherited('overscroll-behavior-y', 'auto', 'contain');
</script>
</body>
</html>
